+++ java/org/lwjgl/opengl/ARBVertexBufferObject.java
@@ -477,7 +477,7 @@ public final class ARBVertexBufferObject {
 	}
 
 	/** Explicit size alternative version of: {@link #glMapBufferARB MapBufferARB} */
-	public static ByteBuffer glMapBufferARB(int target, int access, int length, ByteBuffer old_buffer) {
+	public static ByteBuffer glMapBufferARB(int target, int access, long length, ByteBuffer old_buffer) {
 		long __result = nglMapBufferARB(target, access);
 		return old_buffer != null && __result == memAddress0(old_buffer) && old_buffer.capacity() == length ? old_buffer : memByteBuffer(__result, length);
 	}
